requirements.txt
Flask==2.3.3
Flask-PyMongo==2.3.0
pymongo==4.5.0
python-dotenv==1.0.0
gunicorn==21.2.0
Werkzeug==2.3.7
Jinja2==3.1.2
click==8.1.7
itsdangerous==2.1.2
MarkupSafe==2.1.3
dnspython==2.4.2
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
yt-dlp==2025.9.26
requests==2.31.0
python-dotenv==1.0.0
pysrt==1.1.2
sentence-transformers
elasticsearch==8.11.0
elasticsearch-dsl==8.11.0
uuid==1.30
huggingface_hub
